v1.0.43

+++ application/admin/controller/AgentCancel.php
@@ -67,7 +67,7 @@ class AgentCancel extends Admin
 	}
 
 
-	/**
+/**
 	 * 运营商取消合作确认通过
 	 * @return [type] [description]
 	 */
@@ -76,60 +76,67 @@ class AgentCancel extends Admin
 		// 获取取消合作订单id、获取运营商id、获取取消合作理由、获取运营商名称
 		$data = input('post.');
 		Db::startTrans();
-		// 修改运营商旗下修车厂状态为取消合作后继续做邦保养,小程序不显示。
-		$audit_status = Db::table('cs_shop')->where('aid',$data['aid'])->setField('audit_status',6);
-		if($audit_status !== false){
+		$status = Db::table('ca_apply_cancel')->where('aid',$data['aid'])->setField('status',1);
+		if($status !== false){
 
-			// 修车厂除去用户的油，剩余油品。把所有修车厂都录入到修车厂取消合作表里,给运营商增加授信库存
-			$shop_cancel = $this->shopCancel($data['aid']);
-			if($shop_cancel !==false){
-				// 获取运营商现有库存，入库到运营商取消合作表
-				$agent_ration =Db::table('ca_ration cr')
-								->join('co_bang_cate bc','cr.materiel = bc.id')
-								->where('aid',$data['aid'])->field('aid,materiel,materiel_stock,open_stock,name')->select();
-				$arr['detail']=$agent_ration;
-				$result = Db::table('ca_apply_cancel')->json(['detail'])->insert($arr);
-				if($result){
-					//修改运营商为取消合作。
-					$res = Db::table('ca_agent')->where('aid',$data['aid'])->setField('status',6);
-					if($res !== false){
-						 // 清除运营商的库存
-						 $clear = Db::table('ca_ration')->where('aid',$data['aid'])->delete();
-						 if($clear){
-						 	// 给修车厂发送短信
-						 	$oilSms = $this->oilSms($data['aid'],$data['company'],$data['reason']);
-						 	if($oilSms == '请求成功' || $oilSms == '该运营商旗下没有修车厂'){
-						 		Db::commit();
-								$this->result('',1,'取消合作成功');
-						 	}else{
-						 		Db::rollback();
-								$this->result('',1,$oilSms);
-						 	}
-						 	
-						 }else{
-						 	Db::rollback();
-							$this->result('',0,'取消合作失败');
-						 }
+			// 修改运营商旗下修车厂状态为取消合作后继续做邦保养,小程序不显示。
+			$audit_status = Db::table('cs_shop')->where('aid',$data['aid'])->setField('audit_status',6);
+			if($audit_status !== false){
+
+				// 修车厂除去用户的油，剩余油品。把所有修车厂都录入到修车厂取消合作表里,给运营商增加授信库存
+				$shop_cancel = $this->shopCancel($data['aid']);
+				if($shop_cancel !==false){
+					// 获取运营商现有库存，入库到运营商取消合作表
+					$agent_ration =Db::table('ca_ration cr')
+									->join('co_bang_cate bc','cr.materiel = bc.id')
+									->where('aid',$data['aid'])->field('aid,materiel,materiel_stock,open_stock,name')->select();
+					$arr['detail']=$agent_ration;
+					$result = Db::table('ca_apply_cancel')->json(['detail'])->where('aid',$this->aid)->update($arr);
+					if($result){
+						//修改运营商为取消合作。
+						$res = Db::table('ca_agent')->where('aid',$data['aid'])->setField('status',6);
+						if($res !== false){
+							 // 清除运营商的库存
+							 $clear = Db::table('ca_ration')->where('aid',$data['aid'])->delete();
+							 if($clear){
+							 	// // 给修车厂发送短信
+							 	// $oilSms = $this->oilSms($data['aid'],$data['company'],$data['reason']);
+							 	// if($oilSms == '提交成功' || $oilSms == '该运营商旗下没有修车厂'){
+							 	// 	Db::commit();
+									// $this->result('',1,'取消合作成功');
+							 	// }else{
+							 	// 	Db::rollback();
+									// $this->result('',0,$oilSms);
+							 	// }
+							 	
+							 }else{
+							 	Db::rollback();
+								$this->result('',0,'取消合作失败');
+							 }
+						}else{
+							Db::rollback();
+							$this->result('',0,'修改运营商为取消合作成功');
+						}
 					}else{
 						Db::rollback();
-						$this->result('',0,'修改运营商为取消合作成功');
+						$this->result('',0,'运营商物料详情失败');
 					}
 				}else{
 					Db::rollback();
-					$this->result('',0,'运营商物料详情失败');
+					$this->result('',0,$shop_cancel);
 				}
+
+				
 			}else{
 				Db::rollback();
-				$this->result('',0,$shop_cancel);
+				$this->result('',0,'更改修车厂取消合作失败');
 			}
-
-			
 		}else{
 			Db::rollback();
 			$this->result('',0,'更改修车厂取消合作失败');
 		}
+		
 	}
-
 
 
 
@@ -216,27 +223,52 @@ class AgentCancel extends Admin
 					->select();
 			$user_num[] = $list;
 		}
+		if(!empty($user_num)){
 
-		foreach ($user_num as $ke => $v) {
-			foreach($v as $key=>$value){
-				$arr[] = $value;
-			}
-		}
-
-		// 获取修车厂现有库存，判断留油之后还剩多少库存
-		$ration = $this->shopRation($data);
-		foreach ($ration as $kk => $vv) {
-			foreach($arr as $kkk=>$vvv){
-				if($vv['sid'] == $vvv['sid'] && $vv['materiel'] == $vvv['oil']){
-					// 修改修车厂的库存，为仅做邦保养的库存
-					Db::table('cs_ration')->where('sid',$vvv['sid'])->setField('stock',0);
-					Db::table('cs_ration')->where(['sid'=>$vvv['sid'],'materiel'=>$vvv['remain_times']])->setField('stock',$vvv['remain_times']*$vvv['litre']);
-					$ration[$kk]['stock'] = $vv['stock']-$vvv['remain_times']*$vvv['litre'];
-					
+			foreach ($user_num as $ke => $v) {
+				foreach($v as $key=>$value){
+					$arr[] = $value;
 				}
 			}
+
+			// 获取修车厂现有库存，判断留油之后还剩多少库存
+			$ration = $this->shopRation($data);
+			foreach ($ration as $kk => $vv) {
+				foreach($arr as $kkk=>$vvv){
+					if($vv['sid'] == $vvv['sid'] && $vv['materiel'] == $vvv['oil']){
+						// 修改修车厂的库存，为仅做邦保养的库存
+						Db::table('cs_ration')->where('sid',$vvv['sid'])->setField('stock',0);
+						Db::table('cs_ration')->where(['sid'=>$vvv['sid'],'materiel'=>$vvv['remain_times']])->setField('stock',$vvv['remain_times']*$vvv['litre']);
+						$ration[$kk]['stock'] = $vv['stock']-$vvv['remain_times']*$vvv['litre'];
+						
+					}
+				}
+			}
+			
+		}else{
+			$ration = $this->shopRation($data);
 		}
 		return $ration;
+
+		
+	}
+
+	/**
+	 * 货品交接
+	 * @return [type] [description]
+	 */
+	public function handOver()
+	{
+		$aid = input('post.aid');
+		$list = Db::table('ca_apply_cancel')
+				->where('aid',$aid)
+				->json(['detail'])
+				->find();
+		if($list){
+			$this->result($list,1,'获取交接详情成功');
+		}else{
+			$this->result('',0,'获取交接详情失败');
+		}
 	}
 
 
@@ -290,7 +322,7 @@ class AgentCancel extends Admin
 	{	
 		// $con=$company.'因'.$reason.'停止邦保养服务,您可以'.$content;
         $con = '【'.$company.'】'.'因,【'.$reason.'】停止邦保养服务，您可以更换到附近其他维修厂。';
-		$shop = Db::table('cs_shop')->where('aid',$aid)->field('id,aid,company,leader,reason')->select();
+		$shop = Db::table('cs_shop')->where('aid',$aid)->field('id,aid,company,leader,reason,phone')->select();
 		if($shop){
 			foreach ($shop as $k => $v) {
 				$res = $this->smsVerify($v['phone'],$con);
